+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - test/proxy: add the `Proxy::node_launch_id()` method.
 - logger: add the `format.colorization` parameter to force colors in logs.
 - network: emit the `elfo_network_io_write_waiting_seconds` metric.
+- network: emit the `elfo_message_waiting_time_seconds` metric for all messages.
 
 ### Fixed
 - network: decrease the `elfo_network_rx_flows` metric only in network actors.